+++ CHANGELOG.rst
@@ -2,6 +2,13 @@
 Changelog
 =========
 
+Version 3.0
+===========
+
+* Update HTTP endpoints for circuit execution and results retrieval. `#26 <https://github.com/iqm-finland/iqm-client/pull/26>`_
+* Requires CoCoS 4.0
+
+
 Version 2.2
 ===========